+++ AppView/Controllers/PuchaseMethodController .cs
@@ -19,11 +19,25 @@ namespace AppView.Controllers
             _pu = _dbContext.PurchaseMethods;
             AllRepositories<PurchaseMethod> all = new AllRepositories<PurchaseMethod>(_dbContext, _pu);
             _repos = all;
-        }  
-   
+        }
+
+        private bool CheckUserRole()
+        {
+            var CustomerRole = HttpContext.Session.GetString("UserId");
+            var EmployeeNameSession = HttpContext.Session.GetString("RoleName");
+            var EmployeeName = EmployeeNameSession != null ? EmployeeNameSession.Replace("\"", "") : null;
+            if (CustomerRole != null || EmployeeName != "Quản lý")
+            {
+                return false;
+            }
+            return true;
+        }
         public async Task<IActionResult> GetAllPu()
         {
-
+            if (CheckUserRole() == false)
+            {
+                return RedirectToAction("Forbidden", "Home");
+            }
             string apiUrl = "https://localhost:7036/api/PurchaseMethod/get-pu";
             var httpClient = new HttpClient();
             var response = await httpClient.GetAsync(apiUrl);
@@ -37,6 +51,10 @@ namespace AppView.Controllers
         [HttpGet]
         public async Task<IActionResult> CreatePu()
         {
+            if (CheckUserRole() == false)
+            {
+                return RedirectToAction("Forbidden", "Home");
+            }
             return View();
         }
         [HttpPost]
@@ -51,16 +69,16 @@ namespace AppView.Controllers
 		[HttpGet]
 		public async Task<IActionResult> EditPu(Guid id) // Khi ấn vào Create thì hiển thị View
 		{
-			// Lấy Product từ database dựa theo id truyền vào từ route
-			PurchaseMethod role = _repos.GetAll().FirstOrDefault(c => c.PurchaseMethodID == id);
+            if (CheckUserRole() == false)
+            {
+                return RedirectToAction("Forbidden", "Home");
+            }
+            // Lấy Product từ database dựa theo id truyền vào từ route
+            PurchaseMethod role = _repos.GetAll().FirstOrDefault(c => c.PurchaseMethodID == id);
 			return View(role);
 		}
 		public async Task<IActionResult> EditPu(PurchaseMethod role) // Thực hiện việc Tạo mới
 		{
-			/*var httpClient = new HttpClient();
-			string apiUrl = $"https://localhost:7036/api/Role/update-role?RoleCode={role.RoleCode}&RoleName={role.RoleName}&Status={role.Status}&DateCreated={role.DateCreated}&ColorID={role.RoleID}";
-			var response = await httpClient.PutAsync(apiUrl, null);
-			return RedirectToAction("GetAllRole");*/
 			if (_repos.EditItem(role))
 			{
 				return RedirectToAction("GetAllPu");
@@ -69,7 +87,11 @@ namespace AppView.Controllers
 		}
 		public async Task<IActionResult> DeletePu(Guid id)
 		{
-			var cus = _repos.GetAll().First(c => c.PurchaseMethodID == id);
+            if (CheckUserRole() == false)
+            {
+                return RedirectToAction("Forbidden", "Home");
+            }
+            var cus = _repos.GetAll().First(c => c.PurchaseMethodID == id);
 			var httpClient = new HttpClient();
 			string apiUrl = $"https://localhost:7036/api/PurchaseMethod/delete-pu?id={id}";
 			var response = await httpClient.DeleteAsync(apiUrl);
@@ -77,7 +99,11 @@ namespace AppView.Controllers
 		}
 		public async Task<IActionResult> FindPu(string searchQuery)
 		{
-			var color = _repos.GetAll().Where(c => c.MethodName.ToLower().Contains(searchQuery.ToLower()) );
+            if (CheckUserRole() == false)
+            {
+                return RedirectToAction("Forbidden", "Home");
+            }
+            var color = _repos.GetAll().Where(c => c.MethodName.ToLower().Contains(searchQuery.ToLower()) );
 			return View(color);
 		}
